+++ doc/sample_phrase_annotation.png
