+++ doc/template_config_matlab_spectrograms.ini
