+++ src/cnn_bilstm/curvefit.py
