+++ src/cnn_bilstm/evfuncs.py
