+++ src/cnn_bilstm/koumura_utils.py
