+++ src/cnn_bilstm/model.py
